+++ src/main/php/util/collections/Queue.class.php
@@ -163,10 +163,6 @@ class Queue extends \lang\Object {
    * @return  bool
    */
   public function equals($cmp) {
-    return (
-      $cmp instanceof self && 
-      $this->__generic === $cmp->__generic &&
-      $this->_hash === $cmp->_hash
-    );
+    return $cmp instanceof self && $this->_hash === $cmp->_hash;
   }
 }